classification/requirements.txt
matplotlib
numpy
pandas
scikit-learn
tensorboard
pytorch
torchvision
pillow
opencv
